Refactor plat scripts related to enemy settings
Since enemy values of Health, Damange, Attack
Speed, Attack Impulse, Move Speed are setted
in the script SearchSpace.cs and represents
values related to the Top-down game, these
values need to be ajusted to the platform game.

This commit permits the use of the value
generated and putted in EnemySo object
(that contains enemy settings) from
the evolutive Map-Elite approach and transforms it
in range values setted manually in the Inspector, for
every platform enemy.